Refaktorerer. Legger alle komponenter som kun er relevante for forsiden i en egen mappe ved navn forside
